[Feat] : 요청 받은 문자열을 통해 구글 클라우드 조회
