+++ build.gradle
@@ -20,6 +20,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
     implementation 'org.apache.poi:poi:5.2.0'
@@ -29,6 +30,7 @@ dependencies {
     implementation group: 'org.springdoc', name: 'springdoc-openapi-ui', version: '1.6.4'
     implementation group: 'org.springdoc', name: 'springdoc-openapi-security', version: '1.6.4'
     testImplementation group: 'org.springdoc', name: 'springdoc-openapi-webmvc-core', version: '1.6.4'
+    testImplementation 'com.github.tomakehurst:wiremock-jre8:2.33.0'
 
     implementation group: 'io.jsonwebtoken', name: 'jjwt-api', version: '0.11.2'
     testImplementation 'org.projectlombok:lombok:1.18.26'
