+++ src/main/java/com/alal/backend/config/security/SecurityConfig.java
@@ -81,13 +81,11 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter{
                     .authenticationEntryPoint(new CustomAuthenticationEntryPoint())
                     .and()
                 .authorizeRequests()
-                    .antMatchers("/api/v1/test")
-                        .permitAll()
                     .antMatchers("/", "/error","/favicon.ico", "/**/*.png", "/**/*.gif", "/**/*.svg", "/**/*.jpg", "/**/*.html", "/**/*.css", "/**/*.js")
                         .permitAll()
                     .antMatchers("/swagger", "/swagger-ui.html", "/swagger-ui/**", "/api-docs", "/api-docs/**", "/v3/api-docs/**")
                         .permitAll()
-                    .antMatchers("/login/**","/auth/**", "/oauth2/**")
+                    .antMatchers("/login/**","/auth/**", "/oauth2/**", "/view/**")
                         .permitAll()
                     .antMatchers("/blog/**")
                         .permitAll()
