+++ src/main/java/com/alal/backend/controller/auth/AuthController.java
@@ -144,5 +144,4 @@ public class AuthController {
 
         return ResponseEntity.ok(responseTestToken);
     }
-
 }
